% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{all.equal.key}
\alias{all.equal.key}
\title{An all.equal method for variable wide keys}
\usage{
\method{all.equal}{key}(target, current, ..., check.attributes = FALSE)
}
\arguments{
\item{target}{A wide variable key}

\item{current}{A wide variable key}

\item{...}{Other arguments that are ignored}

\item{check.attributes}{Default FALSE}
}
\description{
Disregards attributes by defaults. Before comparing the two keys,
the values are sorted by \code{"name_new")}.
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}
