% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date.R
\name{as_date}
\alias{as_date}
\title{Parse date strings into Date objects}
\usage{
as_date(x)
}
\arguments{
\item{date_str}{\link{character} vector of date/time strings to parse.}
}
\value{
A \link[lubridate:date_utils]{lubridate::Date} vector corresponding to parsed input. Unparseable entries
are returned as NA.
}
\description{
This function attempts to parse character date strings using
lubridate::parse_date_time with a set of common orders. If a string cannot be
parsed it yields NA in the result.
}
\details{
The parsing orders tried (in no particular priority) are:
"dmy", "mdy", "ymd", "Ymd", "dmY", "mdY", and "Ydm". Parsing is
performed with exact = FALSE to allow flexible separators and
variants supported by lubridate.
}
\examples{
# Single string
as_date("01-02-2003")
# Vector of mixed formats
as_date(c("2003/02/01", "01 Mar 2004", "31-12-1999"))

}
\seealso{
lubridate::parse_date_time
}
