% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{checkValues}
\alias{checkValues}
\title{Compare observed values with the values listed
(presumably from a variable key).}
\usage{
checkValues(x, value_old, xname, diagnostic = FALSE)
}
\arguments{
\item{x}{a variable, either character or factor}

\item{value_old}{a vector of old values for which we are checking}

\item{xname}{character string to use for x's name when printing output}

\item{diagnostic}{prints messages about variables if TRUE}
}
\description{
This is purely diagnostic. It prints warnings in
either of 2 cases. 1)  observed data has values that
are not in the value_old, or 2) that value_old has
values that are not in the data.
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
\keyword{internal}
