% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{dev.create}
\alias{dev.create}
\title{Create a graphics device}
\usage{
dev.create(...)
}
\arguments{
\item{...}{Currently, height and width parameters that would be
suitable with dev.new}
}
\description{
This is a way to create a graphic device on screen that can
display R plots. It is performing the same purpose as R's dev.new,
but it overcomes the limitations of RStudio.  It is needed because
RStudio does not implement fully the functionality of
dev.new. This is suitable for Windows, Linux, and
Macintosh operating systems.
}
\details{
The argument in dev.new named noRStudioGD seems to aimed at same
puroose. But it does not do what I want and documentation is too
sparse.
}
\examples{
\donttest{
dev.create(height = 7, width = 3)
dev.off()
}
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
