% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dms}
\alias{dms}
\title{Delete multiple slashes, replace with one}
\usage{
dms(name)
}
\arguments{
\item{name}{A character string to clean}
}
\description{
Sometimes paths end up with "/too//many//slashes".
While harmless, this is untidy. Clean it up.
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
