% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataPath.R
\name{getDataPath}
\alias{getDataPath}
\title{Give the complete path of the data and cache data on local disk}
\usage{
getDataPath(
  path,
  ...,
  use_cache = TRUE,
  cfg = loadConfig(),
  root = cfg$data[[cfg$data$mode]],
  cache = Sys.getenv("PKG_DATA_CACHE", file.path(dirname(tempdir()),
    utils::packageName()))
)
}
\arguments{
\item{path}{\link{character} representing the path of the data in the database}

\item{...}{\link{character} path elements to add after \code{path}}

\item{use_cache}{\link{logical} download the data and use cache location}

\item{cfg}{a config object. Configuration to use. See \link[fairify]{loadConfig} for details}

\item{root}{\link{character} URL or path of the database root}

\item{cache}{\link{character} path where downloaded files from cloud are cached (See details)}
}
\value{
\link{character} with the path of the file to read either on:
\itemize{
\item the local storage if \code{cfg$data$mode == "local"}
\item the cache directory if \code{cfg$data$mode != "local"} and \code{use_cache == TRUE}
\item the cloud URL if \code{cfg$data$mode != "local"} and \code{use_cache == FALSE}
}
}
\description{
It supports Nextcloud public links in case of using cloud. In this case,
this function can download the file in a temporary
folder and returns the path of the downloaded file.
This functionality is useful for \code{read_excel} from the package \emph{readxl}
(See: https://stackoverflow.com/a/41368947/5300212).
}
\details{
The cache directory is by default located in the system temporary folder in a
subdirectory named as the package name or the value of the environment variable
"PKG_DATA_CACHE" if it is defined.
}
