% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{keyChecker}
\alias{keyChecker}
\title{Checks a variable key for possible errors.}
\usage{
keyChecker(key)
}
\arguments{
\item{key}{variable key object or a file path to a key}
}
\description{
Checks a variable key for possible errors.
}
\author{
Ben Kite <bakite@ku.edu
}
\keyword{internal}
