% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{keyTemplateSPSS}
\alias{keyTemplateSPSS}
\title{Import an SPSS file, create a key representing the numeric -> factor transition}
\usage{
keyTemplateSPSS(dat, long = TRUE)
}
\arguments{
\item{dat}{A character string path to the SPSS file}

\item{long}{TRUE returns a long key, otherwise a wide key}
}
\value{
A variable key (long or wide)
}
\description{
This is a way to keep track of the scores that are used in the SPSS file
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
