% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{keyTemplateStata}
\alias{keyTemplateStata}
\title{Import a Stata (version 12 or lower) file, create a key
representing the numeric -> factor transition}
\usage{
keyTemplateStata(dat, long = TRUE)
}
\arguments{
\item{dat}{A character string path to the Stata file}

\item{long}{TRUE returns a long key, otherwise a wide key}
}
\value{
A variable key (long or wide)
}
\description{
This is a way to keep track of the scores that are used in the Stata file.
It is also an easy way to start a new variable key that makes it convenient
to work on the value_new column with R text functions.
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
