% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kutils-package.R
\docType{package}
\name{kutils-package}
\alias{kutils}
\alias{kutils-package}
\title{kutils: kutils - R package for utils functions}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
