% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Read package configuration}
\usage{
loadConfig(
  userFile = "config.yml",
  pathDefaultCfg = system.file("config.yml", package = utils::packageName()),
  ...
)
}
\arguments{
\item{userFile}{location of the user config YML file (Use NULL or empty character to ignore)}

\item{pathDefaultCfg}{The location of the default configuration (located in "inst/config.yml" of the package by default)}

\item{...}{Other parameters than \code{file} sent to \link[config:get]{config::get} such as \code{config} or \code{value}}
}
\value{
A configuration as it is returned by \link[config:get]{config::get}
}
\description{
Read default package configuration and complete it with eventual user config.
}
\details{
User file configuration (by default the \code{config.yml} file in the working
directory) can overwrite items of the default configuration (See examples).
}
