% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{makeKeylist}
\alias{makeKeylist}
\title{Convert the variable key into a keylist structure for use in
keyApply}
\usage{
makeKeylist(key, sep = c(character = "\\\\|", logical = "\\\\|", integer =
  "\\\\|", factor = "[\\\\|<]", ordered = "[\\\\|<]", numeric = "\\\\|"))
}
\arguments{
\item{key}{A key object or a file name, csv, xlsx or rds.}

\item{sep}{Separator regular expressions}
}
\value{
A list with one element per variable name, along with some
    attributes like class_old and class_new. The class is set as
    well, "keylist".
}
\description{
The return value is a list, with one element per "name_old" to
"name_new" variable combination.  If the key has one old variable
being recoded 6 ways, that begets 6 elements in the resulting
list. Attributes including the classes of the old and new
variables are included.
}
\author{
Paul Johnson
}
\keyword{internal}

