% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semTable.R
\name{markupConvert}
\alias{markupConvert}
\title{Convert marked-up characters to latex, html, or csv}
\usage{
markupConvert(marked, type = c("latex", "html", "csv"), table.float = FALSE,
  longtable = FALSE, caption = NULL, label = NULL, file = NULL, columns)
}
\arguments{
\item{marked}{A character string}

\item{type}{Output type, can be a vector or any one of "latex",
"html", and "csv".}

\item{table.float}{TRUE if you want insertion of '\\begin{table}'}

\item{longtable}{should a tabular or a longtable object be created?}

\item{caption}{A caption to use if either longtable or table is TRUE}

\item{label}{A LaTeX label for cross-references}

\item{file}{A file stub, to which ".tex", ".html", or ".csv" can be added}

\item{columns}{For SEM table, the list of columns objects}
}
\value{
a list of marked up character objects
}
\description{
The conversion key tables are included in the code of the function
}
\author{
Paul Johnson
}
