% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{naLast}
\alias{naLast}
\title{Sort key so that non missing values are first in the
value vector.}
\usage{
naLast(
  key,
  byvar = "name_new",
  valvar = "value_new",
  na.strings = c("\\\\.", "", "\\\\s+", "N/A")
)
}
\arguments{
\item{key}{key, long or wide}

\item{byvar}{Default is "name_new", the column for sorting into blocks.}

\item{valvar}{The value variable on which sorting is to be done, "value_new".}

\item{na.strings}{vector of characters to be treated as NA}
}
\value{
sorted key, with blocks that have the missings last
}
\description{
Sort key so that non missing values are first in the
value vector.
}
\keyword{internal}
