% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCheck.R
\name{print.keycheck}
\alias{print.keycheck}
\title{Print out the result of mergeCheck function.}
\usage{
\method{print}{keycheck}(x, ...)
}
\arguments{
\item{x}{keycheck output from mergeCheck}

\item{...}{Other arguments}
}
\value{
None, side effect if print to screen
}
\description{
This is a placeholder for a more elaborate print method
to be prepared in the future. Please advise us what might
be most helpful.
}
\author{
Paul Johnson
}
