% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{print.keyDiagnostic}
\alias{print.keyDiagnostic}
\title{Print the "changes" component of a keyDiagnostic object}
\usage{
\method{print}{keyDiagnostic}(x, ...)
}
\arguments{
\item{x}{A keyDiagnostic object}

\item{...}{Other arguments passed through to print}
}
\description{
Print the "changes" component of a keyDiagnostic object
}
\author{
Ben Kite <bakite@ku.edu>
}
