% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{print.likert}
\alias{print.likert}
\title{print method for likert tables}
\usage{
\method{print}{likert}(x, ...)
}
\arguments{
\item{x}{likert object, 1st item will be printed}

\item{...}{Arguments passed to print method}
}
\value{
Nothing
}
\description{
Nothing fancy here. \code{cat} is called on first item in list
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
