% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{reverse}
\alias{reverse}
\title{Reverse the levels in a factor}
\usage{
reverse(x, eol = c("Skip", "DNP"))
}
\arguments{
\item{x}{a factor variable}

\item{eol}{values to be kept at the end of the list}
}
\value{
a new factor variable with reversed values
}
\description{
simple literal reversal. No fancy checks on input. If there are
valid levels named "Skip" or "DNP", they will be moved back to the
end of the list.
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
