% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{smartRead}
\alias{smartRead}
\title{read file after deducing file type from suffix.}
\usage{
smartRead(file, ...)
}
\arguments{
\item{file}{name of file to be imported, including path to
file. file name must end in "csv", "xlsx" or "rds"}

\item{...}{additional arguments for read.csv or read.xlsx.}
}
\value{
A data frame or matrix.
}
\description{
read file after deducing file type from suffix.
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
\keyword{internal}

