% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{smartSave}
\alias{smartSave}
\title{save file after deducing type from suffix}
\usage{
smartSave(obj, file, outdir)
}
\arguments{
\item{obj}{an object}

\item{file}{file name. must end in "csv", "xlsx" or "rds"}

\item{outdir}{directory path}
}
\description{
just a convenience for this work, not general purpose
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
\keyword{internal}
