% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{sortStanza}
\alias{sortStanza}
\title{Move missing values to last row in long key block}
\usage{
sortStanza(
  keyblock,
  valvar = "value_new",
  na.strings = c("\\\\.", "", "\\\\s+", "N/A")
)
}
\arguments{
\item{keyblock}{a variable key, or section of rows}

\item{valvar}{Default is "value_new", the column for sorting.}

\item{na.strings}{Stings to be treated as missing, along with R's
NA symbol}
}
\value{
keyblock, row re-arranged with missings last
}
\description{
Receive key stanza and sort the rows so that the missing
values are last in the list. Leaves ordering of other rows
unchanged otherwise.
}
\keyword{internal}
