% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableKey.R
\name{statdatKey}
\alias{statdatKey}
\title{keyFactors: private function that does work for keyTemplateSPSS and
key template Stata}
\usage{
statdatKey(datf, datn, long = TRUE)
}
\arguments{
\item{datf}{Data frame with factors}

\item{datn}{Numeric data frame}

\item{long}{Should the result be a long or wide key}
}
\description{
keyFactors: private function that does work for keyTemplateSPSS and
key template Stata
}
